# contracts/openapi/v1.locked.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SyncRunSummaryPayload": {
        "additionalProperties": true,
        "description": "What a HealthSave client says about a run it just closed.\n\nReceipts are written per HTTP batch, so a run that had nothing to send left\nno trace on the server and ``GET /api/v2/sync/runs/latest`` kept answering\nwith the previous run. This body is the run's existence proof: one small\nPUT per completed run, idempotent on the run id. It carries counts and\nmetric NAMES only \u2014 never a health value.\n\n``extra='allow'`` keeps the shape forward-compatible (a newer client may add\nkeys an older server ignores); the enumerated fields are validated\ndeterministically and a bad value is a 422, never a 500.",
        "properties": {
          "client_app_version": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client App Version"
          },
          "client_platform": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Platform"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "delivery": {
            "enum": [
              "none",
              "foreground",
              "background_queued"
            ],
            "title": "Delivery",
            "type": "string"
          },
          "error_class": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Class"
          },
          "intent": {
            "anyOf": [
              {
                "enum": [
                  "latest_changes",
                  "backfill",
                  "date_range"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent"
          },
          "metrics_checked": {
            "items": {
              "type": "string"
            },
            "maxItems": 4096,
            "title": "Metrics Checked",
            "type": "array"
          },
          "metrics_with_changes": {
            "items": {
              "type": "string"
            },
            "maxItems": 4096,
            "title": "Metrics With Changes",
            "type": "array"
          },
          "outcome": {
            "enum": [
              "completed",
              "failed"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "records_sent": {
            "minimum": 0.0,
            "title": "Records Sent",
            "type": "integer"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "enum": [
              1
            ],
            "title": "Schema Version",
            "type": "integer"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "trigger": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger"
          }
        },
        "required": [
          "outcome",
          "delivery",
          "records_sent"
        ],
        "title": "SyncRunSummaryPayload",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "HealthSave Observatory",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/apple/batch": {
      "post": {
        "description": "Receive a metric batch from HealthSave iOS app.\n\nExpected payload:\n{\n    \"metric\": \"heart_rate\",\n    \"batch_index\": 0,\n    \"total_batches\": 1,\n    \"samples\": [\n        {\"date\": \"2024-01-15T10:30:00Z\", \"qty\": 72, \"source\": \"Apple Watch\"},\n        ...\n    ]\n}",
        "operationId": "apple_batch_api_apple_batch_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apple Batch"
      }
    },
...
    "/api/apple/status": {
      "get": {
        "description": "Return record counts so the iOS app knows what's synced.",
        "operationId": "apple_status_api_apple_status_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apple Status"
      }
    },
...
    "/api/v2/apple/batch": {
      "post": {
        "description": "Receive a v2 Apple HealthKit batch.\n\nWire contract: ``POST /api/v2/apple/batch`` with ``schema_version=2``.\nAccepts the same idempotency / receipt headers as v1 plus an optional\n``X-HealthSave-Schema-Version: 2`` (advisory \u2014 the body schema_version\nis the source of truth).\n\nResponse: same delivery-receipt shape as v1 with an additive\n``wire_schema_version: 2`` and a ``deletions`` block carrying the\nper-table supersede counts.",
        "operationId": "v2_apple_batch_api_v2_apple_batch_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "V2 Apple Batch"
      }
    },
...
    "/api/v2/sync/anomalies": {
      "get": {
        "description": "Detect suspicious sync behavior visible from server receipts.",
        "operationId": "sync_anomalies_api_v2_sync_anomalies_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Sync Anomalies Api V2 Sync Anomalies Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Anomalies"
      }
    },
    "/api/v2/sync/coverage": {
      "get": {
        "description": "Return metric-level receipt and destination sample coverage.",
        "operationId": "sync_coverage_api_v2_sync_coverage_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Sync Coverage Api V2 Sync Coverage Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Coverage"
      }
    },
    "/api/v2/sync/runs/latest": {
      "get": {
        "description": "Summarize the most recently observed HealthSave sync run.",
        "operationId": "latest_sync_run_api_v2_sync_runs_latest_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Latest Sync Run Api V2 Sync Runs Latest Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Latest Sync Run"
      }
    },
    "/api/v2/sync/runs/{sync_run_id}": {
      "get": {
        "description": "Return the delivery receipt summary for one HealthSave sync run.",
        "operationId": "sync_run_api_v2_sync_runs__sync_run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "sync_run_id",
            "required": true,
            "schema": {
              "title": "Sync Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Sync Run Api V2 Sync Runs  Sync Run Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Run"
      }
    },
    "/api/v2/sync/runs/{sync_run_id}/summary": {
      "put": {
        "description": "Record (or replace) the client's closing summary for one sync run.\n\nOptional for third-party servers, like every ``/api/v2/sync/*`` route: a\nclient treats 404/405 here as \"not supported\" and carries on. When present,\nit is what lets ``GET /api/v2/sync/runs/latest`` report a run that sent\nnothing as the latest run instead of silently repeating the previous one.",
        "operationId": "put_sync_run_summary_api_v2_sync_runs__sync_run_id__summary_put",
        "parameters": [
          {
            "in": "path",
            "name": "sync_run_id",
            "required": true,
            "schema": {
              "title": "Sync Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Api-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRunSummaryPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Put Sync Run Summary Api V2 Sync Runs  Sync Run Id  Summary Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Sync Run Summary"
      }
    },
...
  }
}